+++ lib/restclient.rb
@@ -1,14 +1,10 @@
 require 'uri'
 require 'zlib'
 require 'stringio'
+require 'net/https'
 
-begin
-  require 'net/https'
-rescue LoadError => e
-  raise e unless RUBY_PLATFORM =~ /linux/
-  raise LoadError, "no such file to load -- net/https. Try running apt-get install libopenssl-ruby"
-end
-
+require File.dirname(__FILE__) + '/restclient/version'
+require File.dirname(__FILE__) + '/restclient/platform'
 require File.dirname(__FILE__) + '/restclient/exceptions'
 require File.dirname(__FILE__) + '/restclient/request'
 require File.dirname(__FILE__) + '/restclient/abstract_response'
@@ -16,7 +12,7 @@ require File.dirname(__FILE__) + '/restclient/response'
 require File.dirname(__FILE__) + '/restclient/raw_response'
 require File.dirname(__FILE__) + '/restclient/resource'
 require File.dirname(__FILE__) + '/restclient/payload'
-require File.dirname(__FILE__) + '/restclient/net_http_ext'
+require File.dirname(__FILE__) + '/restclient/windows'
 
 # This module's static methods are the entry point for using the REST client.
 #
@@ -101,12 +97,6 @@ module RestClient
   # You can also configure logging by the environment variable RESTCLIENT_LOG.
   def self.log= log
     @@log = create_log log
-  end
-
-  def self.version
-    version_path = File.dirname(__FILE__) + "/../VERSION"
-    return File.read(version_path).chomp if File.file?(version_path)
-    "0.0.0"
   end
 
   # Create a log that respond to << like a logger