+++ spec/unit/restclient_spec.rb
@@ -70,4 +70,10 @@ describe RestClient do
     end
   end
 
+  describe 'version' do
+    it 'has a version ~> 1.7.0.alpha' do
+      ver = Gem::Version.new(RestClient.version)
+      Gem::Requirement.new('~> 1.7.0.alpha').should be_satisfied_by(ver)
+    end
+  end
 end